+++ amp.Database/DataModel/VersionInfo.cs
@@ -24,13 +24,8 @@ SOFTWARE.
 */
 #endregion
 
-using System;
-using System.Collections.Generic;
 using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
 
 namespace amp.Database.DataModel;
