+++ amp.EtoForms/FormMain.Methods.cs
@@ -29,6 +29,8 @@ using System.Diagnostics.CodeAnalysis;
 using System.Reflection;
 using amp.DataAccessLayer;
 using amp.DataAccessLayer.DtoClasses;
+using amp.Database;
+using amp.Database.Migration;
 using amp.Database.QueryHelpers;
 using amp.EtoForms.Dialogs;
 using amp.EtoForms.ExtensionClasses;
@@ -36,14 +38,17 @@ using amp.EtoForms.Forms.Enumerations;
 using amp.EtoForms.Properties;
 using amp.EtoForms.Settings.Enumerations;
 using amp.EtoForms.Utilities;
+using amp.Playback;
 using amp.Playback.Classes;
 using amp.Shared.Constants;
 using amp.Shared.Localization;
 using Eto.Drawing;
 using Eto.Forms;
 using EtoForms.Controls.Custom.Helpers;
+using EtoForms.Controls.Custom.UserIdle;
 using EtoForms.Controls.Custom.Utilities;
 using EtoForms.SpectrumVisualizer;
+using Microsoft.Data.Sqlite;
 using Microsoft.EntityFrameworkCore;
 using VPKSoft.Utils.Common.UpdateCheck;
 
@@ -555,5 +560,25 @@ SOFTWARE.
         }
 
         return current;
+    }
+
+    private void SuspendBackgroundTasks()
+    {
+        idleChecker.Dispose();
+        AttachDetachPlaybackManagerEvents(false);
+        playbackManager.Dispose();
+        context.Dispose();
+        SqliteConnection.ClearAllPools();
+    }
+
+    private void ResumeBackgroundTasks()
+    {
+        context = new AmpContext();
+        SoftwareMigration.RunSoftwareMigration(context);
+        idleChecker = new UserIdleChecker(this);
+        playbackManager = new PlaybackManager<AudioTrack, AlbumTrack, Album>(GetNextAudioTrackFunc, GetTrackById,
+            Globals.Settings.PlaybackRetryCount);
+        AttachDetachPlaybackManagerEvents(true);
+        playbackManager.ManagerStopped = false;
     }
 }
