+++ amp.EtoForms/FormMain.cs
@@ -34,7 +34,6 @@ using amp.Shared.Localization;
 using Eto.Drawing;
 using Eto.Forms;
 using EtoForms.Controls.Custom.UserIdle;
-using Microsoft.Data.Sqlite;
 using VPKSoft.Utils.Common.EventArgs;
 using VPKSoft.Utils.Common.Interfaces;
 using Album = amp.DataAccessLayer.DtoClasses.Album;
@@ -111,27 +110,6 @@ public partial class FormMain : Form, IExceptionReporter
         AssignEventListeners();
         CreateMenu();
     }
-
-    private void SuspendBackgroundTasks()
-    {
-        idleChecker.Dispose();
-        AttachDetachPlaybackManagerEvents(false);
-        playbackManager.Dispose();
-        context.Dispose();
-        SqliteConnection.ClearAllPools();
-    }
-
-    private void ResumeBackgroundTasks()
-    {
-        context = new AmpContext();
-        SoftwareMigration.RunSoftwareMigration(context);
-        idleChecker = new UserIdleChecker(this);
-        playbackManager = new PlaybackManager<AudioTrack, AlbumTrack, Album>(GetNextAudioTrackFunc, GetTrackById,
-            Globals.Settings.PlaybackRetryCount);
-        AttachDetachPlaybackManagerEvents(true);
-        playbackManager.ManagerStopped = false;
-    }
-
 
     private void TestStuff_Executed(object? sender, EventArgs e)
     {
